+++ public/index.php
@@ -33,6 +33,18 @@
 
 $content = ob_get_clean();
 
-$layout = c('main', [ 'content' => $content ], LAYOUTS_DIR);
+$styles = [
+
+];
+
+$scripts = [
+
+];
+
+$layout = c('main', [
+    'content' => $content,
+    'styles'  => $styles,
+    'scripts' => $scripts,
+], LAYOUTS_DIR);
 
 echo $layout;